Sem3 Web Project/requirements.txt
langchain
openai
PyPDF2
faiss-cpu
huggingface-hub
python-dotenv
sentence-transformers